# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trip Planner API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trips/{trip_id}/ai/ideas": {
      "post": {
        "tags": [
          "suggestions"
        ],
        "summary": "Ask For Ideas",
        "description": "Queue a Claude run that suggests things to do. Poll the run, then read the trip's suggestions.",
        "operationId": "ask_for_ideas",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeasIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/suggestions": {
      "get": {
        "tags": [
          "suggestions"
        ],
        "summary": "List Suggestions",
        "description": "The trip's suggestions, newest run first. Without `status`, the new and added ones.",
        "operationId": "list_suggestions",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "new",
                      "added",
                      "dismissed"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestionOut"
                  },
                  "title": "Response List Suggestions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suggestions/{suggestion_id}": {
      "patch": {
        "tags": [
          "suggestions"
        ],
        "summary": "Update Suggestion",
        "description": "Dismiss a suggestion, or bring a dismissed one back.",
        "operationId": "update_suggestion",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suggestions/{suggestion_id}/add": {
      "post": {
        "tags": [
          "suggestions"
        ],
        "summary": "Add Suggestion",
        "description": "Add the suggestion to the itinerary: planned at its day and time, or as an idea with `as_idea`.",
        "operationId": "add_suggestion",
        "parameters": [
          {
            "name": "suggestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suggestion Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AddSuggestionIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "enum": [
              "sights",
              "museum",
              "food",
              "nature",
              "nightlife",
              "shopping",
              "travel",
              "other"
            ],
            "title": "Category"
          },
          "status": {
            "type": "string",
            "enum": [
              "idea",
              "planned",
              "booked"
            ],
            "title": "Status"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "place_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Provider"
          },
          "place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          },
          "place_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Data"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trip_id",
          "day",
          "start_time",
          "end_time",
          "title",
          "category",
          "status",
          "location_name",
          "address",
          "lat",
          "lon",
          "url",
          "notes",
          "place_provider",
          "place_id",
          "place_data",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "ActivityOut"
      },
...
      "AddSuggestionIn": {
        "properties": {
          "as_idea": {
            "type": "boolean",
            "title": "As Idea",
            "default": false
          }
        },
        "type": "object",
        "title": "AddSuggestionIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdeasIn": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "brainstorm",
              "surprise"
            ],
            "title": "Mode",
            "default": "brainstorm"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          }
        },
        "type": "object",
        "title": "IdeasIn"
      },
...
      "RunOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "routine_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Id"
          },
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "flight_api",
              "flight_agent",
              "research_agent",
              "itinerary_agent",
              "lodging_agent"
            ],
            "title": "Kind"
          },
          "trigger": {
            "type": "string",
            "enum": [
              "schedule",
              "manual",
              "catch_up"
            ],
            "title": "Trigger"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "partial",
              "failed",
              "timed_out",
              "cancelled",
              "interrupted"
            ],
            "title": "Status"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queued At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "accepted_count": {
            "type": "integer",
            "title": "Accepted Count"
          },
          "rejected_count": {
            "type": "integer",
            "title": "Rejected Count"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "cost_usd_est": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd Est"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          }
        },
        "type": "object",
        "required": [
          "id",
          "routine_id",
          "trip_id",
          "kind",
          "trigger",
          "status",
          "params",
          "queued_at",
          "started_at",
          "finished_at",
          "summary",
          "error",
          "accepted_count",
          "rejected_count",
          "input_tokens",
          "output_tokens",
          "cost_usd_est",
          "cancel_requested"
        ],
        "title": "RunOut"
      },
...
      "SuggestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "brainstorm",
              "surprise"
            ],
            "title": "Mode"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "enum": [
              "sights",
              "museum",
              "food",
              "nature",
              "nightlife",
              "shopping",
              "travel",
              "other"
            ],
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "why": {
            "type": "string",
            "title": "Why"
          },
          "timing_note": {
            "type": "string",
            "title": "Timing Note"
          },
          "day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "added",
              "dismissed"
            ],
            "title": "Status"
          },
          "activity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "trip_id",
          "run_id",
          "mode",
          "title",
          "category",
          "description",
          "why",
          "timing_note",
          "day",
          "start_time",
          "duration_min",
          "location_name",
          "url",
          "sources",
          "status",
          "activity_id",
          "created_at",
          "end_time"
        ],
        "title": "SuggestionOut"
      },
      "SuggestionUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "new",
              "dismissed"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SuggestionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}